PSG-1844: move check_metadata and staging of input samples to common so that they can be reused by other pipelines
