+++ scripts/common/check_metadata.py
